requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
crispy-bootstrap5==2025.4
Django==5.0.6
django-ckeditor==6.7.2
django-ckeditor-5==0.2.18
django-crispy-forms==2.4
django-js-asset==3.1.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==22.0.0
idna==3.10
packaging==24.1
pillow==11.2.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.7.0
django-widget-tweaks==1.5.0
